Minor tweaks: Create platform components.
