+++ shared/core/base/src/androidMain/kotlin/com/thomaskioko/tvmaniac/base/BasePlatformComponent.kt
@@ -1,5 +1,7 @@
 package com.thomaskioko.tvmaniac.base
 
+import com.thomaskioko.tvmaniac.base.model.AppCoroutineDispatchers
+import com.thomaskioko.tvmaniac.base.model.AppCoroutineScope
 import com.thomaskioko.tvmaniac.base.scope.ApplicationScope
 import com.thomaskioko.tvmaniac.base.util.AndroidAppUtils
 import com.thomaskioko.tvmaniac.base.util.AndroidDateFormatter
@@ -9,9 +11,30 @@ import com.thomaskioko.tvmaniac.base.util.AppUtils
 import com.thomaskioko.tvmaniac.base.util.DateFormatter
 import com.thomaskioko.tvmaniac.base.util.ExceptionHandler
 import com.thomaskioko.tvmaniac.base.util.FormatterUtil
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Job
 import me.tatarka.inject.annotations.Provides
 
-interface BasePlatformComponent {
+actual interface BasePlatformComponent {
+
+    @ApplicationScope
+    @Provides
+    fun provideCoroutineDispatchers(): AppCoroutineDispatchers = AppCoroutineDispatchers(
+        io = Dispatchers.IO,
+        computation = Dispatchers.Default,
+        main = Dispatchers.Main,
+    )
+
+    @ApplicationScope
+    @Provides
+    fun provideCoroutineScope(
+        dispatchers: AppCoroutineDispatchers
+    ): AppCoroutineScope = AppCoroutineScope(
+        default = CoroutineScope(Job() + dispatchers.computation),
+        io = CoroutineScope(Job() + dispatchers.io),
+        main = CoroutineScope(Job() + dispatchers.main),
+    )
 
     @ApplicationScope
     @Provides
