+++ shared/core/base/src/commonMain/kotlin/com/thomaskioko/tvmaniac/base/BasePlatformComponent.kt
@@ -1,0 +1,3 @@
+package com.thomaskioko.tvmaniac.base
+
+expect interface BasePlatformComponent
