+++ shared/core/base/src/iosMain/kotlin/com/thomaskioko/tvmaniac/base/scope/NsQueueCoroutineScope.kt
@@ -1,0 +1,15 @@
+package com.thomaskioko.tvmaniac.base.scope
+
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Job
+import kotlin.coroutines.CoroutineContext
+
+class NsQueueCoroutineScope : CoroutineScope {
+
+    private val coroutineDispatcher: CoroutineDispatcher = applicationNsQueueDispatcher
+    private val job = Job()
+
+    override val coroutineContext: CoroutineContext
+        get() = job + coroutineDispatcher
+}
