+++ shared/core/base/src/iosMain/kotlin/com/thomaskioko/tvmaniac/base/scope/NsQueueDispatcher.kt
@@ -1,0 +1,52 @@
+package com.thomaskioko.tvmaniac.base.scope
+
+import kotlinx.coroutines.*
+import platform.darwin.*
+import kotlin.coroutines.CoroutineContext
+
+@OptIn(InternalCoroutinesApi::class)
+val applicationNsQueueDispatcher: CoroutineDispatcher = NsQueueDispatcher(dispatch_get_main_queue())
+
+@InternalCoroutinesApi
+internal class NsQueueDispatcher(private val dispatchQueue: dispatch_queue_t) : CoroutineDispatcher(), Delay {
+
+    override fun dispatch(context: CoroutineContext, block: Runnable) {
+        dispatch_async(dispatchQueue) {
+            block.run()
+        }
+    }
+
+    override fun scheduleResumeAfterDelay(timeMillis: Long, continuation: CancellableContinuation<Unit>) {
+        dispatch_after(dispatch_time(DISPATCH_TIME_NOW, timeMillis * 1_000_000), dispatchQueue) {
+            try {
+                with(continuation) {
+                    resumeUndispatched(Unit)
+                }
+            } catch (err: Throwable) {
+                throw err
+            }
+        }
+    }
+
+    override fun invokeOnTimeout(timeMillis: Long, block: Runnable, context: CoroutineContext): DisposableHandle {
+        val handle = object : DisposableHandle {
+            var disposed = false
+                private set
+
+            override fun dispose() {
+                disposed = true
+            }
+        }
+        dispatch_after(dispatch_time(DISPATCH_TIME_NOW, timeMillis * 1_000_000), dispatchQueue) {
+            try {
+                if (!handle.disposed) {
+                    block.run()
+                }
+            } catch (err: Throwable) {
+                throw err
+            }
+        }
+
+        return handle
+    }
+}
