+++ shared/core/networkutil/src/androidMain/kotlin/com/thomaskioko/tvmaniac/core/networkutil/NetworkUtilComponent.kt
@@ -3,7 +3,7 @@ package com.thomaskioko.tvmaniac.core.networkutil
 import com.thomaskioko.tvmaniac.base.scope.ApplicationScope
 import me.tatarka.inject.annotations.Provides
 
-interface NetworkUtilComponent {
+actual interface NetworkUtilComponent {
 
     @ApplicationScope
     @Provides
