+++ shared/core/networkutil/src/iosMain/kotlin/com/thomaskioko/tvmaniac/core/networkutil/NetworkUtilComponent.kt
@@ -1,0 +1,3 @@
+package com.thomaskioko.tvmaniac.core.networkutil
+
+actual interface NetworkUtilComponent
