--- shared/core/util/src/iosMain/kotlin/com/thomaskioko/tvmaniac/core/util/scope/NsQueueCoroutineScope.kt
@@ -1,15 +1,0 @@
-package com.thomaskioko.tvmaniac.core.util.scope
-
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Job
-import kotlin.coroutines.CoroutineContext
-
-class NsQueueCoroutineScope : CoroutineScope {
-
-    private val coroutineDispatcher: CoroutineDispatcher = applicationNsQueueDispatcher
-    private val job = Job()
-
-    override val coroutineContext: CoroutineContext
-        get() = job + coroutineDispatcher
-}
