+++ shared/data/database/src/commonMain/kotlin/com.thomaskioko.tvmaniac.db/DatabaseComponent.kt
@@ -1,0 +1,3 @@
+package com.thomaskioko.tvmaniac.db
+
+expect interface DatabaseComponent
