+++ shared/data/datastore/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/datastore/implementation/DataStorePlatformComponent.kt
@@ -1,0 +1,3 @@
+package com.thomaskioko.tvmaniac.datastore.implementation
+
+expect interface DataStorePlatformComponent
